Auth0 Rule for adding group claim to token
Trello: https://trello.com/c/QJTNs9QM

The rule adds auth0 groups and github teams to a namespaced groups claim
in a user's id token.

Group claims are classified as custom claims and must be namespaced to be oidc
compliant. See:
https://auth0.com/docs/api-auth/tutorials/adoption/scope-custom-claims

These claims can later be consumed/queried by applications/services.

I added this rule for assigning roles to groups via role-bindings with the k8s api, but this is not the only use case.