Challenge for 2FA every 8 hours + Refactoring
Before this change we were challenging for 2FA at every login.
This was alright but as we want to have a more streamlined user experience
and enable silent SSO this is not a problem.

The rule now check when was last time a user logged in (by getting the full
Auth0 profile which contains `last_login`) and only challenge if he/she
didn't login in the last 8 hours.

NOTE: From empiric experience it seems like the user goes through this
      rule twice. When user goes through rule the first time their `last_login`
      field is updated even if they didn't enter the 2FA code yet (unfortunately).

      This is the reason why if user logged in less than 5 seconds I consider
      that as a non-login and re-enforce 2FA (otherwise they will not be
      challenged). I also tested that if user waits before clicking on
      "login" button they will still be challenged for 2FA code.