Challenge Control Panel users for 2FA
Background:

kubernetes, CP-API and CP-UI use the same Auth0 client (as JWT tokens are
passed from UI down to UI).

Old behaviour:

The rule ignored `kubectl-oidc` client as the kubectl is a command line
utility and we don't challenge for 2FA in that context.

As a side effect of this, we also don't challenge users for 2FA when
logging in CP.

New behaviour:

The rule is now not challenging for 2FA when client is using a refresh token
(as they already authenticated and this is also non-interactive).