Replace `context.connection` with something more reliable
We found that a rule was not working as expected as it was skipped even
when it shouldn't have. This was caused by the following statement:

    // Apply rule only if logging in via GitHub
    if (context.connection !== 'github') {
      // ...
    }

The problem was that sometimes `context.connection` is not set causing
the rule to be skipped when it shouldn't.

I've updated similar logic here to avoid this kind of problem.

Rules now look at user identities. Our Auth0 users have only one identity
so `context.connection` is replaced with `user.identities[0].connection`.

RULES UPDATED
=============
- `add-group-claim-to-token`: This only needs to run when logging using
  GH via certain clients
- `auth0-authorization-extension`: Only runs when connection type is `email`.
  I updated the logic and also replaced unnecessary logic to determine
  connection (`context.connection || user.identities[0].connection`) with
  a constant (`email`) as the connection will always be email in this rule.
- `Multifactor-Google-Authenticator-Do-Not-Rename`: Only applied to certain
  connections
- `whitelist-google-domains`: Only apply when logging in via Google

TICKET
======

https://trello.com/c/ldYtid7e/775-1-auth0-rules-dont-use-contextconnection